Big changes
A large update, including implementation of stamina bar, a run cycle animation for Mr. Grant, and preparations to background objects. The stamina bar and Mr. Grant's run cycle are dependent upon the speed variable. In future updates, Mr. Grant should be stabilized in the middle of the screen and objects should move relative to him. The background update is the start towards this framework.